# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API: каталог конструкций, заявки на дендрохронологический анализ. Без авторизации доступны только GET каталога; с JWT — методы пользователя; модератор — создание конструкций и завершение заявок.",
        "title": "Dendrochronology API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/constructions": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "constructions"
                ],
                "summary": "Список конструкций",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Поиск по названию",
                        "name": "query",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/serializer.ConstructionJSON"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "multipart/form-data; только для is_moderator.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "constructions"
                ],
                "summary": "Создать конструкцию",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название",
                        "name": "construction_title",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Use-life",
                        "name": "use_life",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Описание",
                        "name": "description",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Изображение",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Видео",
                        "name": "video",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/serializer.ConstructionJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/constructions/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "constructions"
                ],
                "summary": "Конструкция по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializer.ConstructionJSON"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "serializer.ConstructionJSON": {
            "type": "object",
            "properties": {
                "construction_title": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "is_delete": {
                    "type": "boolean"
                },
                "use_life": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Передавайте JWT: \"Bearer \\\u003ctoken\\\u003e\" (тот же токен, что в cookie auth_token после входа).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}